Create README.md
The viruses like SARS-CoV2 are mutating everytime and making themselves resistant to the existing drugs and antibodies. So we came with a solution to this problem using machine learning, we predict the binding affinities of antibody and RBD of virus complex before the virus gets mutated, so that scientists can know whether a mutation will be resistant to the existing antibodies  or not in the future and if they are making themselves resistant to the antibody scientists can make changes in the antibody accordingly to make them non resistant.